feat: useLogout 훅 추가
 - useLogout 훅 추가
 - 일부 코드 리팩토링
