+++ src/features/profile/LogoutConfirmButton.tsx
@@ -1,20 +1,8 @@
 import { Button } from '@/components/ui/button';
-import { userKeys } from '@/hooks/queries/user/userQueryKeys';
-import { logoutAuth } from '@/services/apis/authApi';
-import { useMutation, useQueryClient } from '@tanstack/react-query';
+import { useLogout } from '@/hooks/queries/user/useLogout';
 
 export default function LogoutConfirmButton() {
-  const queryClient = useQueryClient();
-  const { mutate } = useMutation({
-    mutationFn: logoutAuth,
-    onSuccess() {
-      queryClient.invalidateQueries({ queryKey: userKeys.all });
-      window.location.href = '/';
-    },
-    onError(err) {
-      console.log(err);
-    },
-  });
+  const { mutate } = useLogout();
 
   const onLogoutClick = () => {
     mutate();
