+++ src/hooks/useUser.tsx
@@ -1,4 +1,4 @@
-import { currentUser } from '@/services/apis/userApi';
+import { authUserApi } from '@/services/apis/userApi';
 import { useQuery } from '@tanstack/react-query';
 import { useEffect } from 'react';
 import { useNavigate } from 'react-router-dom';
@@ -9,7 +9,7 @@ type UserState = 'userOnly' | 'nonUserOnly';
 export default function useUser(userState?: UserState) {
   const { data, error } = useQuery({
     queryKey: userKeys.all,
-    queryFn: () => currentUser.getUser(),
+    queryFn: () => authUserApi.getUser(),
   });
 
   const nav = useNavigate();
